requirements.txt
flask
mysql-connector-python
werkzeug
gunicorn
flask-socketio
eventlet
